OSX, C, refact: move network func in net.c
