+++ MANIFEST.in
@@ -97,6 +97,8 @@ include psutil/arch/osx/disk.c
 include psutil/arch/osx/disk.h
 include psutil/arch/osx/mem.c
 include psutil/arch/osx/mem.h
+include psutil/arch/osx/net.c
+include psutil/arch/osx/net.h
 include psutil/arch/osx/process_info.c
 include psutil/arch/osx/process_info.h
 include psutil/arch/osx/sensors.c
