+++ psutil/_psutil_osx.c
@@ -17,7 +17,6 @@
 #include <sys/proc_info.h>
 #include <netinet/tcp_fsm.h>
 #include <arpa/inet.h>
-#include <net/if_dl.h>
 #include <pwd.h>
 #include <unistd.h>
 #include <mach/mach.h>
@@ -29,6 +28,7 @@
 #include "_psutil_posix.h"
 #include "arch/osx/cpu.h"
 #include "arch/osx/disk.h"
+#include "arch/osx/net.h"
 #include "arch/osx/mem.h"
 #include "arch/osx/process_info.h"
 #include "arch/osx/sensors.h"
@@ -878,92 +878,6 @@ psutil_proc_num_fds(PyObject *self, PyObject *args) {
 
     free(fds_pointer);
     return Py_BuildValue("i", num_fds);
-}
-
-
-/*
- * Return a Python list of named tuples with overall network I/O information
- */
-static PyObject *
-psutil_net_io_counters(PyObject *self, PyObject *args) {
-    char *buf = NULL, *lim, *next;
-    struct if_msghdr *ifm;
-    int mib[6];
-    mib[0] = CTL_NET;          // networking subsystem
-    mib[1] = PF_ROUTE;         // type of information
-    mib[2] = 0;                // protocol (IPPROTO_xxx)
-    mib[3] = 0;                // address family
-    mib[4] = NET_RT_IFLIST2;   // operation
-    mib[5] = 0;
-    size_t len;
-    PyObject *py_ifc_info = NULL;
-    PyObject *py_retdict = PyDict_New();
-
-    if (py_retdict == NULL)
-        return NULL;
-
-    if (sysctl(mib, 6, NULL, &len, NULL, 0) < 0) {
-        PyErr_SetFromErrno(PyExc_OSError);
-        goto error;
-    }
-
-    buf = malloc(len);
-    if (buf == NULL) {
-        PyErr_NoMemory();
-        goto error;
-    }
-
-    if (sysctl(mib, 6, buf, &len, NULL, 0) < 0) {
-        PyErr_SetFromErrno(PyExc_OSError);
-        goto error;
-    }
-
-    lim = buf + len;
-
-    for (next = buf; next < lim; ) {
-        ifm = (struct if_msghdr *)next;
-        next += ifm->ifm_msglen;
-
-        if (ifm->ifm_type == RTM_IFINFO2) {
-            py_ifc_info = NULL;
-            struct if_msghdr2 *if2m = (struct if_msghdr2 *)ifm;
-            struct sockaddr_dl *sdl = (struct sockaddr_dl *)(if2m + 1);
-            char ifc_name[32];
-
-            strncpy(ifc_name, sdl->sdl_data, sdl->sdl_nlen);
-            ifc_name[sdl->sdl_nlen] = 0;
-
-            py_ifc_info = Py_BuildValue(
-                "(KKKKKKKi)",
-                if2m->ifm_data.ifi_obytes,
-                if2m->ifm_data.ifi_ibytes,
-                if2m->ifm_data.ifi_opackets,
-                if2m->ifm_data.ifi_ipackets,
-                if2m->ifm_data.ifi_ierrors,
-                if2m->ifm_data.ifi_oerrors,
-                if2m->ifm_data.ifi_iqdrops,
-                0);  // dropout not supported
-
-            if (!py_ifc_info)
-                goto error;
-            if (PyDict_SetItemString(py_retdict, ifc_name, py_ifc_info))
-                goto error;
-            Py_CLEAR(py_ifc_info);
-        }
-        else {
-            continue;
-        }
-    }
-
-    free(buf);
-    return py_retdict;
-
-error:
-    Py_XDECREF(py_ifc_info);
-    Py_DECREF(py_retdict);
-    if (buf != NULL)
-        free(buf);
-    return NULL;
 }
